--- a/zio-query/shared/src/main/scala/zio/query/internal/BlockedRequests.scala
+++ b/zio-query/shared/src/main/scala/zio/query/internal/BlockedRequests.scala
@@ -16,10 +16,10 @@
 
 package zio.query.internal
 
-import zio.query.internal.BlockedRequests._
 import zio.query._
+import zio.query.internal.BlockedRequests._
 import zio.stacktracer.TracingImplicits.disableAutoTrace
-import zio.{Chunk, Exit, FiberId, Promise, Trace, UIO, Unsafe, ZEnvironment, ZIO}
+import zio.{Chunk, Exit, Promise, Trace, UIO, Unsafe, ZEnvironment, ZIO}
 
 import scala.annotation.tailrec
 import scala.collection.mutable