@kyri-petrou 2024-04-03
> Does it improve the caliban benchmarks too?

@ghostdogpr Doubtful, the optimizations are for datasource-backed queries and we only use `ZQuery.succeed` in the Caliban benchmarks. However, I think we should add a param in Caliban in the execution config to size the Cache. Default value of 16 seems way too low