@kyri-petrou 2024-04-03
> Maybe try to run the caliban test suite as well to catch anything

I don't think we have anything in Caliban's test suites that uses datasources. I'll test it against some projects at $WORK when we have a snapshot version of zio-query and caliban which heavily rely on ZQueries